move the assignee display-name helper into utils
