+++ src/ui/primitives/Avatar.ts
@@ -1,21 +1,5 @@
-import { parseLinktext, setTooltip } from 'obsidian'
-import { stringToColor } from '../../utils'
-
-export function displayName(raw: string): string {
-  const trimmed = raw.trim()
-  const m = trimmed.match(/^\[\[([^\]]+)\]\]$/)
-  if (!m) return trimmed
-  const inner = m[1]
-  const pipe = inner.indexOf('|')
-  if (pipe >= 0) {
-    const alias = inner.slice(pipe + 1).trim()
-    if (alias) return alias
-  }
-  const target = pipe >= 0 ? inner.slice(0, pipe) : inner
-  const { path } = parseLinktext(target)
-  const base = path.split('/').pop() ?? path
-  return (base.endsWith('.md') ? base.slice(0, -3) : base).trim()
-}
+import { setTooltip } from 'obsidian'
+import { displayName, stringToColor } from '../../utils'
 
 function initialsFor(name: string): string {
   const parts = name.trim().split(/\s+/).filter(Boolean)
