+++ src/views/ProjectOverviewView.ts
@@ -3,8 +3,8 @@ import type PMPlugin from '../main'
 import type { Project, ResolvedProjectConfig, Task } from '../types'
 import { collectAllTags, flattenTasks, totalLoggedHours, type ProjectRef } from '../store'
 import { Temporal, parsePlainDate, today } from '../dates'
-import { dateUrgency, formatDateLong, isTerminalStatus, safeAsync, truncateTitle } from '../utils'
-import { Avatar, displayName } from '../ui/primitives/Avatar'
+import { dateUrgency, displayName, formatDateLong, isTerminalStatus, safeAsync, truncateTitle } from '../utils'
+import { Avatar } from '../ui/primitives/Avatar'
 import { Chip } from '../ui/primitives/Chip'
 import { EmptyState } from '../ui/primitives/EmptyState'
 import { ProgressBar } from '../ui/primitives/ProgressBar'
